xinetzone/pyproject.toml
[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core"]

[project]
authors = [{name = "xinetzone", email = "735613050@qq.com"}]
classifiers = ["License :: OSI Approved :: Apache Software License"]
dynamic = ["version", "description"]
license = {file = "LICENSE"}
name = "labelme"
readme = "../README.md"
requires-python = ">=3.9"

dependencies = [
    "gdown",
    "imgviz>=1.7.5",
    "matplotlib",
    "natsort>=7.1.0",
    "numpy",
    "onnxruntime>=1.14.1,!=1.16.0",
    "osam>=0.2.2",
    "Pillow>=2.8",
    "PyYAML",
    "qtpy!=1.11.2",
    "scikit-image",
    "termcolor",
]

maintainers = [
  {name = "xinetzone", email = "735613050@qq.com"},
]

[project.urls]
Home = "https://github.com/xinetzone/labelme"

[project.optional-dependencies]
doc = [
  "sphinx",
  "docutils",
]

dev = [
  "pdm",
  "gymnasium",
  "PyGithub[integrations]",
  "bytecode",
  "numpy",
  "toml",
  "fsspec",
  "pyarrow",
  "invoke",
  "openpyxl",
  "rich[jupyter]"
]

gui = [
  "PySide6",
]

[tool.flit.module]
name = "labelme"

[tool.flit.sdist]
# include = ["doc/"]
exclude = ['doc', '.git', '.github', 'output']

[pypi.extra]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
# verify_ssl = false
